-creaing osx executable
-
