+++ extractInfo.py
@@ -1,72 +1,13 @@
 import json
 import datetime
-import csv 
+import csv
 import sys
-import os 
-
-# class definition for object representing a  trello card 
-class cardData(object):
-    def __init__(self,jsob,List,idtoName):
-        """ Class containing most relevant information from a trello card
-        Attributes: 
-            List (str): Specifies the list the card belongs to. 
-            name (str): Card Title 
-            due (datetime): date time object describing the due date for the card 
-            desc (str):  string containing the cards discription. 
-            url (str): string contiaing short url to this card 
-            member(str):  string representing all the people assigned to that card 
-            labels (str): string containing trello cards label information 
-            completion (bool): boolean representing whether the task was completed 
-
-        """
-        self.List = List 
-        self.name = jsob['name'] 
-        self.due = jsob['due']
-        self.completion = jsob['dueComplete']
-        self.desc  = jsob['desc']
-        self.url = jsob['shortUrl']
-        self.labels= ", "
-        self.member = ", " 
-
-        if jsob['idMembers']: #accounts for the lack of assigned people to cards :
-            self.member = self.member.join(list(map(lambda x: idtoName[x],jsob['idMembers']) ) )
-        else: 
-            self.member = "N/A"
-
-        if jsob['due']: 
-            self.due= datetime.datetime.strptime(jsob['due'],'%Y-%m-%dT%H:%M:%S.%fZ')
-        else:
-            self.due = datetime.datetime(2021,7,1)# if due date exist i set july 1,2021 as the "due" date 
-        if jsob['labels']:
-            self.labels = self.labels.join( list(map(lambda x: x['name'],jsob['labels'])) )
-        else: 
-            self.labels = "N/A" 
-
-    #string representation of the object 
-    def __str__(self):
-        return   self.name
-    def __repr__(self): 
-        return  self.name 
-
-    #functions used whenver the object is to be compared i.e( x<y and such  ) 
-    def __lt__(self,other):
-        return self.due < other.due 
-    def __le__(self,other):
-        return self.due<= other.due 
-    def __eq__(self,other): 
-        return self.due == other.due 
-    def __ne__(self,other):
-        return self.due != other.due 
-    def __gt__(self,other):
-        return self.due > other.due 
-    def __ge__(self,other):
-        return self.due >= otehr.due 
-
-    #simple function to  represent object as a single row of a csv 
-    def toCSV(self):
-        """" create list of strings to represent card in csv file """ 
-        return [self.due.date(),self.labels,self.name,self.List,self.desc,self.member,self.url]
-
+import os
+from cardData import cardData 
+# class definition for object representing a  trello card
+"""
+    Script used to get jsonfile informaiton 
+"""
 def mapIDmembertoName(parsed_data):
     """ creates dictionary mapping user id to their actual name """
     idtoName = dict()
@@ -98,20 +39,19 @@ def flatCardStack(cardStack):
     data = list()
     for k in cardStack.keys():
         for e in cardStack[k]: 
-            data.append(e) 
-    return sorted(data) 
-
-
-if __name__== '__main__':
-    print('Attempting to open:  %s ' %sys.argv[1])
+            data.append(e)
+    return sorted(filter(lambda x: not( x.completion)  , data)) 
+ 
+def main(s): 
+    print('Attempting to open:  %s ' %s)
     #open and load data 
-    data= open(sys.argv[1])
+    data= open(s)
     parsed_data = json.load(data) 
     # create dictionaries  to map from trello id to board and from trello id to member 
     idToName = mapListIdtoName(parsed_data)
     memberMap = mapIDmembertoName(parsed_data) 
     # extract each idividual boards cards seperately then stack them together 
-    things=  flatCardStack(consumeList(parsed_data,idToName,memberMap) ) 
+    things = flatCardStack(consumeList(parsed_data,idToName,memberMap))
     # writting each cards infomation one row at a time 
     with open('output.csv','w') as f:
         writer = csv.writer(f)
